requirements.txt
streamlit
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
numpy
pdfplumber
python-docx
